Move files to correct folder
